refactor(agent): nest vibe page as vibe::setup module
Restructure vibe_setup into vibe/setup (filename-based module pattern):
vmux_agent::vibe::setup with page/event submodules. Drop the source
comments per repo convention. build.rs now tracks both the vibe module
root and directory so page edits rerun the web bundle (addresses review).
