+++ crates/vmux_agent/src/vibe.rs
@@ -1,0 +1,1 @@
+pub mod setup;
