+++ crates/vmux_agent/src/vibe/setup/event.rs
@@ -1,5 +1,3 @@
-/// Emitted by the vibe-setup page (button click) → handled on the Bevy side to spawn a terminal and
-/// run the Vibe install script.
 #[derive(
     Clone,
     Debug,
